y-a-foot-services/src/main/java/net/andresbustamante/yafoot/services/impl/PlayerSearchServiceImpl.java
package net.andresbustamante.yafoot.services.impl;

import net.andresbustamante.yafoot.dao.PlayerDAO;
import net.andresbustamante.yafoot.exceptions.DatabaseException;
import net.andresbustamante.yafoot.model.UserContext;
import net.andresbustamante.yafoot.model.Joueur;
import net.andresbustamante.yafoot.services.PlayerSearchService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

/**
 * @author andresbustamante
 */
@Service
public class PlayerSearchServiceImpl implements PlayerSearchService {

    @Autowired
    private PlayerDAO playerDAO;

    @Override
    public Joueur findPlayerByEmail(String email, UserContext userContext) throws DatabaseException {
        return playerDAO.findPlayerByEmail(email);
    }
}
